Some more refactor to be able to access helper functions from other integration tests
