+++ packages/framework-integration-tests/integration/helper/auth-helper.ts
@@ -1,0 +1,5 @@
+import { internet } from 'faker'
+
+export const createPassword = (): string => {
+  return `${internet.password(8)}Passw0rd!`
+}
