+++ packages/framework-integration-tests/integration/providers/aws/functionality/auth.integration.ts
@@ -6,7 +6,6 @@ import {
   confirmUser,
   createUser,
   deleteUser,
-  createPassword,
   getUserAuthInformation,
   DisconnectableApolloClient,
   countSubscriptionsItems,
@@ -22,6 +21,7 @@ import fetch from 'cross-fetch'
 import { ApolloClient } from 'apollo-client'
 import { NormalizedCacheObject } from 'apollo-cache-inmemory'
 import { waitForIt } from '../../../helper/sleep'
+import { createPassword } from '../../../helper/auth-helper'
 
 chai.use(require('chai-as-promised'))
