+++ packages/framework-integration-tests/integration/providers/aws/nuke/setup.ts
@@ -1,10 +1,10 @@
-import { overrideWithBoosterLocalDependencies } from '../../../helper/depsHelper'
+import { overrideWithBoosterLocalDependencies } from '../../../helper/deps-helper'
 import { nuke } from '../deploy'
 import { setEnv, checkConfigAnd } from '../utils'
 // Imported from another package to avoid duplication
 // It is OK-ish, since integration tests are always run in the context of the whole monorepo
 import { createSandboxProject, removeSandboxProject } from '../../../../../cli/src/common/sandbox'
-import { sandboxPathFor } from '../../../helper/fileHelper'
+import { sandboxPathFor } from '../../../helper/file-helper'
 
 before(async () => {
   await setEnv()
