+++ packages/framework-integration-tests/integration/providers/local/setup.ts
@@ -1,10 +1,10 @@
 import { start } from './utils'
 import { sleep } from '../../helper/sleep'
 import { ChildProcess } from 'child_process'
-import { removeFolders, sandboxPathFor } from '../../helper/fileHelper'
-import { overrideWithBoosterLocalDependencies } from '../../helper/depsHelper'
+import { removeFolders, sandboxPathFor } from '../../helper/file-helper'
+import { overrideWithBoosterLocalDependencies } from '../../helper/deps-helper'
 import { sandboxName } from './constants'
-import { runCommand } from '../../helper/runCommand'
+import { runCommand } from '../../helper/run-command'
 // Imported from another package to avoid duplication
 // It is OK-ish, since integration tests are always run in the context of the whole monorepo
 import { createSandboxProject } from '../../../../cli/src/common/sandbox'
