+++ packages/framework-integration-tests/integration/providers/local/utils.ts
@@ -4,7 +4,7 @@ import { InMemoryCache, NormalizedCacheObject } from 'apollo-cache-inmemory'
 import { HttpLink } from 'apollo-link-http'
 import { LOCAL_PROVIDER_HOST } from './constants'
 import gql from 'graphql-tag'
-import { runCommand } from '../../helper/runCommand'
+import { runCommand } from '../../helper/run-command'
 import path = require('path')
 import { ChildProcess } from 'child_process'
